Move TestHiveCli to hooks.test_hive
