+++ tests/providers/apache/hive/operators/test_hive.py
@@ -67,31 +67,6 @@ class TestHiveEnvironment(unittest.TestCase):
         """
 
 
-class TestHiveCli(unittest.TestCase):
-
-    def setUp(self):
-        self.nondefault_schema = "nondefault"
-        os.environ["AIRFLOW__CORE__SECURITY"] = "kerberos"
-
-    def tearDown(self):
-        del os.environ["AIRFLOW__CORE__SECURITY"]
-
-    def test_get_proxy_user_value(self):
-        from airflow.providers.apache.hive.hooks.hive import HiveCliHook
-
-        hook = HiveCliHook()
-        returner = mock.MagicMock()
-        returner.extra_dejson = {'proxy_user': 'a_user_proxy'}
-        hook.use_beeline = True
-        hook.conn = returner
-
-        # Run
-        result = hook._prepare_cli_cmd()
-
-        # Verify
-        self.assertIn('hive.server2.proxy.user=a_user_proxy', result[2])
-
-
 class HiveOperatorConfigTest(TestHiveEnvironment):
 
     def test_hive_airflow_default_config_queue(self):
